+++ src/main/java/com/assessment/hotelbooking/service/AuthenticationService.java
@@ -1,0 +1,10 @@
+package com.assessment.hotelbooking.service;
+
+import com.assessment.hotelbooking.service.dto.LoginRequest;
+import com.assessment.hotelbooking.service.dto.LoginResponse;
+
+public interface AuthenticationService {
+
+  LoginResponse login(LoginRequest loginRequest);
+
+}